Update on "[flex_attention] adds support for low precision K/V inputs in compiled mode with GPU"
Allow Keys and Values to be in lower precision than Queries for memory efficiency.
K and V are automatically upcast to Q's dtype in the kernel before matmul operations.
In eager mode or compiled mode with CPU, error will be thrown with mixed dtypes.

cc voznesenskym penguinwu EikanWang jgong5 Guobing-Chen XiaobingSuper zhuhaozhe blzheng wenzhe-nrv jiayisunx ipiszy kadeng muchulee8 amjames chauhang aakhundov coconutruben jataylo Chillee drisspg yanboliang BoyuanFeng

[ghstack-poisoned]

+++ test/inductor/test_flex_flash.py
@@ -368,9 +368,9 @@ class TestFlexFlash(InductorTestCase):
         """Ensure flash attention rejects mixed dtypes (e.g., fp32 Q with fp16 K/V)"""
         B, H, S, D = 2, 8, 512, 64
 
-        query = torch.randn(B, H, S, D, dtype=torch.float32, device=device)
-        key = torch.randn(B, H, S, D, dtype=dtype, device=device)
-        value = torch.randn(B, H, S, D, dtype=dtype, device=device)
+        query = torch.randn(B, H, S, D, dtype=torch.bfloat16, device=device)
+        key = torch.randn(B, H, S, D, dtype=dtype, device=device).to(torch.float8_e4m3fn)
+        value = torch.randn(B, H, S, D, dtype=dtype, device=device).to(torch.float8_e4m3fn)
 
         compiled_fn = torch.compile(flex_attention, fullgraph=True)
